pk记录查询显示完成

+++ capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/pojo/StudentExpansion.java
@@ -80,7 +80,7 @@ public class StudentExpansion extends Model<StudentExpansion> {
     /**
      * 最大排行
      */
-    private Integer ranking;
+    private Long ranking;
 
     @Override
     protected Serializable pkVal() {
